Moving tests to specific namespace
