+++ NuGetTest/test/Lombiq.OSOCE.NuGet.Tests.UI/Tests/ModuleTests/BehaviorDataTableTests.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Lombiq.OSOCE.NuGet.Tests.UI.Tests;
+namespace Lombiq.OSOCE.NuGet.Tests.UI.Tests.ModuleTests;
 
 public class BehaviorDataTableTests : UITestBase
 {
